k8s_deployer/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
google-auth==2.16.2
idna==3.4
kubernetes==26.1.0
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.15
websocket-client==1.5.1
psycopg2-binary==2.9.5
SQLAlchemy==2.0.4
jinja2==3.1.2
docker==5.0.3
attrs==21.2.0
cryptography==41.0.2
